chore: storage 파일 분리
